4 Changed returning types from <Bool, Never> to <Void, PersistentStorageError>
