+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version: 5.8
+// swift-tools-version: 5.7
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "PersistentStorage",
     platforms: [
-        .iOS(.v15), .macOS(.v10_15), .tvOS(.v11), .watchOS(.v4)
+        .iOS(.v15), .macOS(.v11), .tvOS(.v13), .watchOS(.v6)
     ],
     products: [
         .library(
